backend/services/calendar-integration-service/requirements.txt
flask
flask-cors
google-api-python-client
python-dateutil
pyyaml